+++ releasenotes.md
@@ -2,6 +2,7 @@
 
 ### DOOM Retro v5.8.1
 
+* *DOOM Retro* now checks for a newer version at startup and notifies the player if one is available.
 * *DOOM Retro* is now built using v17.14.18 of [*Microsoft Visual Studio Community 2022*](https://visualstudio.microsoft.com/vs/community/).
 * Minor changes have been made to text that is output to the console.
 * The `r_levelbrightness` CVAR has been renamed `r_extralighting`.